Move Options to reduce line size
